requirements.txt
sentence-transformers
faiss-cpu
transformers
requests
json
torch
numpy
python-dotenv
